Fix navigation close icon alignment.
Fix useRootPaddingAwareAlignments by using an updated schema
which has the correct location for this setting.

As of this writing, this blog post does not contain the
correct example code for this setting. The setting must be
inside the settings object and not at the root:
https://make.wordpress.org/themes/2022/09/07/full-width-blocks-and-root-padding-in-wordpress-6-1/

The schema is correct and was fixed here:
https://github.com/WordPress/gutenberg/issues/43624

The navigation block uses this setting to correctly render
the close icon:
https://github.com/WordPress/gutenberg/pull/43576